pyproject.toml
[project]
name = "iclr-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=2.4.0",
    "openreview-py>=1.57.0",
    "pymongo>=4.16.0",
    "requests>=2.32.5",
    "typer>=0.23.1",
]
